requirements.txt
fastapi>=0.110.0,<1.0.0
uvicorn[standard]>=0.30.0,<0.33.0
numpy>=1.26.0
joblib>=1.3.0
pydantic>=2.6.0
httpx>=0.27.0
anyio>=4.4.0
pyyaml>=6.0.2
sentence-transformers>=2.7.0
chromadb>=0.4.24
pandas>=2.2.2
scikit-learn>=1.4.0
xgboost>=2.1.0
matplotlib>=3.9.0
pytest>=8.3.1
pytest-asyncio>=0.23.7
sumolib>=1.19.0
traci>=1.19.0
torch
opencv-python
ultralytics
norfair
stable-baselines3[extra]
gym==0.26.2
tensorboard
structlog
flake8
mypy
black
jupyterlab
